requirements.txt
gradio
fastapi
requests
python-jose[cryptography]
uvicorn[standard]